Setup multi-threading and prepared for io and ui threads.
This patch prepares the UI thread and IO thread to receive data. My
main goal in this patch was to properly shutdown the program when it
was did not have a valid parameter as the first argument.

There was a bit of restructure done as well. The EventSocket class now
should properly be able to pass data between threads.