Massive restructure of threading and code base (#4)
This patch revamped threading, task management and cleaned up many files. It introduced a much cleaner way for threads to be managed. Right now threads are passed around using a shared_ptr, however I am not very happy with this and i'll likely change, in the mean time this is how it'll be done.

I also added some debugging and logging support. For ease of use you can use TEST() when testing code and optionally pass in data as format arguments TEST("%s %s", "foo", "bar"). It will output line number and file number for you as well.