Fixed bug in ThreadManager possibly crashing and properly blocking in UI thread
This patch fixes a bug where condition_variable can sometimes spontainously wakeup
this patch ensures that the data is available by the time it does.

Also properly waiting for events in UI Thread. Before it was always blocking and
yielding the thread when nothing to process, now it should properly block on the
conditional variable.